server/src/requirements.txt
aiohttp==4.0.0a0
aiomysql==0.0.20
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.6
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
idna==2.8
more-itertools==7.0.0
multidict==4.5.2
pluggy==0.11.0
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.2
pytest==4.5.0
pytest-aiohttp==0.3.0
python-dotenv==0.10.2
six==1.12.0
SQLAlchemy==1.3.3
wcwidth==0.1.7
yarl==1.3.0
